Revert "refactor: 최상위 패키지를 domain.* -> me.nalab.api.survey.domain.* 로 변경한다"
This reverts commit e2049b1a0d778c3f960d8ad19fa22c47bc8a65f4.
