+++ src/main/kotlin/domain/feedback/Bookmark.kt
@@ -1,4 +1,4 @@
-package me.nalab.api.survey.domain.feedback
+package domain.feedback
 
 import me.nalab.core.time.TimeUtil
 import java.time.Instant
@@ -8,7 +8,7 @@ import javax.persistence.Embeddable
 @Embeddable
 class Bookmark(
     @Column(name = "is_bookmarked", nullable = false)
-    var isBookmarked: Boolean = BOOKMARK_DEFAULT_STATE,
+    var isBookmarked: Boolean = domain.feedback.Bookmark.Companion.BOOKMARK_DEFAULT_STATE,
 
     @Column(name = "bookmarked_at", columnDefinition = "TIMESTAMP(6)", nullable = false)
     var bookmarkedAt: Instant,
@@ -17,7 +17,7 @@ class Bookmark(
     companion object {
         private const val BOOKMARK_DEFAULT_STATE = false
 
-        fun impossible(): Bookmark =
-            Bookmark(bookmarkedAt = TimeUtil.toInstant())
+        fun impossible(): domain.feedback.Bookmark =
+            domain.feedback.Bookmark(bookmarkedAt = TimeUtil.toInstant())
     }
 }
