+++ src/main/kotlin/domain/feedback/ChoiceFormQuestionFeedback.kt
@@ -1,4 +1,4 @@
-package me.nalab.api.survey.domain.feedback
+package domain.feedback
 
 import javax.persistence.*
 
@@ -7,11 +7,11 @@ class ChoiceFormQuestionFeedback(
     id: Long,
     formQuestionId: Long,
     isRead: Boolean = false,
-    bookmark: Bookmark = Bookmark.Companion.impossible(),
-    feedback: Feedback,
+    bookmark: domain.feedback.Bookmark = domain.feedback.Bookmark.Companion.impossible(),
+    feedback: domain.feedback.Feedback,
 
     @ElementCollection
     @Column(name = "selects")
     @CollectionTable(name = "selects", joinColumns = [JoinColumn(name = "form_feedback_id")])
     private val selectedChoiceIds: MutableSet<Long>,
-) : FormQuestionFeedbackable(id, formQuestionId, isRead, bookmark, feedback)
+) : domain.feedback.FormQuestionFeedbackable(id, formQuestionId, isRead, bookmark, feedback)
