+++ src/main/kotlin/domain/feedback/Feedback.kt
@@ -1,4 +1,4 @@
-package me.nalab.api.survey.domain.feedback
+package domain.feedback
 
 import me.nalab.core.data.common.TimeBaseEntity
 import javax.persistence.*
@@ -18,17 +18,17 @@ class Feedback(
         fetch = FetchType.LAZY,
         cascade = [CascadeType.PERSIST, CascadeType.MERGE]
     )
-    val questionFeedbacks: MutableList<FormQuestionFeedbackable> = mutableListOf(),
+    val questionFeedbacks: MutableList<domain.feedback.FormQuestionFeedbackable> = mutableListOf(),
 
     @Column(name = "is_read", nullable = false)
     private var isRead: Boolean = false,
 
     @JoinColumn(name = "reviewer_id", nullable = false)
     @OneToOne(fetch = FetchType.LAZY, cascade = [CascadeType.PERSIST, CascadeType.MERGE])
-    val reviewer: Reviewer,
-) : Comparable<Feedback>, TimeBaseEntity() {
+    val reviewer: domain.feedback.Reviewer,
+) : Comparable<domain.feedback.Feedback>, TimeBaseEntity() {
 
-    override fun compareTo(other: Feedback): Int {
+    override fun compareTo(other: domain.feedback.Feedback): Int {
         if (updatedAt.isAfter(other.updatedAt)) {
             return -1
         }
