+++ src/main/kotlin/domain/survey/QuestionType.kt
@@ -1,4 +1,4 @@
-package me.nalab.api.survey.domain.survey
+package domain.survey
 
 enum class QuestionType {
     CHOICE,
